Waterfall plot of the CatBoost Model
We can observe that our model has good accuracy since all deciles have an agreement between predicted probability and actual probability. The model predicted that the average loss rate for customers in decile 1 is 91%, while, in fact, 94% of customers lost in decile 1. Similarly, the model predicted that the average probability of churn is 63% for customers in decile 2, while the actual likelihood of churn is 64% in this decile. The explanation continues with the rest of the deciles.